Swap order of JSON Decoding
The POST from Travis is encoded with `application/x-www-form-urlencoded` and Sinatra correctly handles decoding this when you access `request.body.read`.  However, this decodes to one key value pair: `payload` and the JSON string of the data.  To properly decode the POST you need to extract the value of `payload` **THEN** parse that as JSON.  The existing version reverses the order of these operations, and I don't believe it will work.